+++ CHANGELOG.rst
@@ -67,7 +67,8 @@ Adding ingestion support (from cgap-portal as initial guide):
   to ``https://raw.githubusercontent.com/The-Sequence-Ontology/SO-Ontologies/master/Ontology_Files/so.owl``.
   and from ``https://www.ebi.ac.uk/efo/efo.owl`` to ``https://github.com/EBISPOT/efo/releases/download/current/efo.owl"``.
 * Updated ``groupfinder`` in ``encoded/authorization.py`` to include ``INGESTION`` in ``localname``
-  check for return of ``group.admin``.
+* Added ``encoded/schemas/file_other.json`` (totally new).
+* Updated ``encoded/types/file.py`` with new ``FileOther`` type.
 
 
 5.1.3